requirements.txt
# ── Core Agent Framework ──────────────────────────────────────────────────────
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0          # LLM provider (OpenAI GPT-4)
langchain-anthropic>=0.2.0       # LLM provider (Anthropic Claude) – optional
langgraph>=0.2.0                 # Agent graph orchestration

# ── MCP (Model Context Protocol) Client ───────────────────────────────────────
mcp>=1.0.0                       # Python MCP SDK to connect to kite-mcp-server
httpx>=0.27.0                    # Async HTTP for MCP HTTP transport
httpx-sse>=0.4.0                 # SSE support for MCP streaming

# ── Zerodha / Indian Market Data ──────────────────────────────────────────────
yfinance>=0.2.40                 # Yahoo Finance – free EOD/historical data (.NS/.BO)
pandas>=2.2.0                    # Data manipulation for financial time series
numpy>=1.26.0                    # Numerical computations

# ── Web Scraping (Quarterly Results) ──────────────────────────────────────────
beautifulsoup4>=4.12.0           # HTML parsing for Screener.in / BSE / NSE
requests>=2.32.0                 # HTTP requests for scraping
lxml>=5.2.0                      # Faster HTML parser backend for BeautifulSoup
fake-useragent>=1.5.0            # Rotate user-agents to avoid scraping blocks

# ── News ──────────────────────────────────────────────────────────────────────
newsapi-python>=0.2.7            # NewsAPI.org client – free tier 100 req/day

# ── Configuration & Validation ────────────────────────────────────────────────
pydantic>=2.7.0                  # Data models and schema validation
pydantic-settings>=2.3.0        # Settings management from .env file
python-dotenv>=1.0.0             # Load .env into environment

# ── Output & CLI ──────────────────────────────────────────────────────────────
rich>=13.7.0                     # Beautiful terminal output / progress bars
typer>=0.12.0                    # CLI framework for on-demand commands
